Move unit tests out of functional test directory
