+++ tests/functional/test_incidents.py
@@ -104,11 +104,3 @@ class TestIncidentsCountEndpoint:
         res = testapp.get('/incidents/count/?fields=leoka_felony')
         for row in res.json:
             assert 'leoka_felony' in row
-
-class TestIncidentsUnit:
-    def test_incidents_list(self):
-        from crime_data.resources.incidents import IncidentsList
-        assert IncidentsList()
-    def test_incidents_count(self):
-        from crime_data.resources.incidents import IncidentsCount
-        assert IncidentsCount()
