+++ tests/unit/test_incidents.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+
+class TestIncidentsUnit:
+    def test_incidents_list(self):
+        from crime_data.resources.incidents import IncidentsList
+        assert IncidentsList()
+
+    def test_incidents_count(self):
+        from crime_data.resources.incidents import IncidentsCount
+        assert IncidentsCount()
